# api-server/src/app/Http/Controllers/Api/TodoController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Api\BaseController as BaseController;
use App\Http\Requests\Todo\TodoStoreRequest;
use App\Services\TodoService;
use Illuminate\Http\Request;

class TodoController extends BaseController
{
  private $service;

  public function __construct(
    TodoService $service
  ) {
    $this->service = $service;
  }

  public function index()
  {
    $todos = $this->service->getTodos();
    $success['todos'] = $todos;

    return $this->sendResponse($success, 'Todos fetched successfully.');
  }

  public function store(TodoStoreRequest $request)
  {
    $todo = $this->service->storeTodo(
      $request->user_id,
      $request->title,
      $request->content,
    );
    $success['todo'] = $todo;
    return $this->sendResponse($success, 'Todo created successfully.');
  }
}